redesign full project

+++ server/controllers/catalogController.js
@@ -1,19 +1,46 @@
 import { Catalog, Product } from "../model/model.js"
+import { deleteDriveFiles, deleteLocalFiles, postDriveFiles } from "../services/googleDriveServices.js"
+import "dotenv"
+
+const uploadCatalogImg = process.env.GOOGLE_DRIVE_CATALOG_PHOTO
 
 class CatalogController {
     async create(req, res) {
+        const { name } = req.body
+        const files = req.files
+        const userId = req.user.id
+
+        if (req.user.role !== "ADMIN") {
+            await deleteLocalFiles()
+            return res.status(403).json({ message: "Доступа запрещен" })
+        }
+
         try {
-            const { name } = req.body
 
             if (!name) {
                 return res.status(400).json({ message: "Не все поля заполнены" })
             }
 
-            const catalog = await Catalog.create({ name })
+            if (files.length === 0) {
+                await deleteLocalFiles()
+                return res.status(406).json({ message: "Не все поля заполнены" })
+            }
+
+            let imageUrls = await postDriveFiles(files, uploadCatalogImg)
+
+            if (imageUrls.length === 0) {
+                await deleteLocalFiles()
+                return res.status(500).json({ message: "Ошибка сервера" })
+            }
+
+            const catalog = await Catalog.create({ userId, name, img: imageUrls })
+
+            await deleteLocalFiles()
 
             return res.status(201).json({ catalog, message: "Каталог успешно создан" })
         } catch (err) {
             console.error(err)
+            await deleteLocalFiles()
             return res.status(500).json({ message: "Ошибка сервера" })
         }
     }
@@ -29,9 +56,32 @@ class CatalogController {
     }
 
     async updateCatalog(req, res) {
+        const { id } = req.params
+        const { name, existingImg } = req.body
+        const files = req.files
+
+        if (req.user.role !== "ADMIN") {
+            await deleteLocalFiles()
+            return res.status(403).json({ message: "Доступа запрещен" })
+        }
+
         try {
-            const { id } = req.params
-            const { name } = req.body
+
+            const existingImages = existingImg ? JSON.parse(existingImg) : []
+            const newFilesCount = files ? files.length : 0
+            const totalImgCount = existingImages.length + newFilesCount
+
+            let catalog = await Catalog.findByPk(id)
+
+            if (totalImgCount > 1) {
+                await deleteLocalFiles()
+                return res.status(400).json({ message: "Максимум 1 фото" })
+            }
+
+            if ((!existingImages || existingImages.length === 0) && (!files || files.length === 0)) {
+                await deleteLocalFiles()
+                return res.status(401).json({ message: "Не все поля заполнены" })
+            }
 
             if (!id) {
                 return res.status(404).json({ message: "Такого каталога не существует" })
@@ -41,9 +91,41 @@ class CatalogController {
                 return res.status(406).json({ message: "Не все поля заполнены" })
             }
 
-            await Catalog.update({ name }, { where: { id } })
+            let imageUrls = catalog.img
 
-            return res.status(203).json({ message: "Каталог успешно обновлен" })
+            const imgToDelete = imageUrls.filter(url => !existingImages.includes(url))
+
+            if (imgToDelete.length > 0) {
+                try {
+                    await deleteDriveFiles(imageUrls)
+                } catch (err) {
+                    console.error("Ошибка при удалении файлов:", err)
+                    await deleteLocalFiles()
+                    return res.status(500).json({ message: "Ошибка при обновлении изображений" })
+                }
+            }
+
+            if (existingImg) {
+                imageUrls = existingImages
+            }
+            let newImgUrls = []
+            if (files && files.length > 0) {
+                try {
+                    newImgUrls = await postDriveFiles(files, uploadCatalogImg)
+                } catch (err) {
+                    console.error("Ошибка загрузки файлов:", err)
+                    await deleteLocalFiles()
+                    return res.status(500).json({ message: "Ошибка при загрузке новых изображений" })
+                }
+            }
+
+            const allImg = [...imageUrls, ...newImgUrls]
+
+            await Catalog.update({ name, img: allImg }, { where: { id } })
+
+            await deleteLocalFiles()
+
+            return res.status(200).json({ message: "Каталог успешно обновлен" })
         } catch (err) {
             console.error(err)
             return res.status(500).json({ message: "Ошибка сервера" })
@@ -51,14 +133,39 @@ class CatalogController {
     }
 
     async deleteCatalog(req, res) {
+        const { id } = req.params
+        const userId = req.user.id
+
+        if (req.user.role !== "ADMIN") {
+            await deleteLocalFiles()
+            return res.status(403).json({ message: "Доступа запрещен" })
+        }
+
         try {
-            const { id } = req.params
 
             if (!id) {
                 return res.status(404).json({ message: "Такого каталога не существует" })
             }
 
-            await Catalog.destroy({ where: { id } })
+            const catalog = await Catalog.findByPk(id, {
+                include: [{
+                    model: Product
+                }]
+            })
+
+            const allFile = [
+                ...(catalog.img || []),
+                ...(catalog.Product?.flatMap(product => product.img) || [])
+            ]
+
+            await Promise.all([
+                deleteDriveFiles(allFile),
+                deleteLocalFiles(),
+
+                Product.destroy({ where: { catalogId: id }, individualHooks: true })
+            ])
+
+            await Catalog.destroy({ where: { id }, userId })
 
             return res.status(200).json({ message: "Каталог успешно удален" })
         } catch (err) {
